# SPA/src/app/core/models/enum-model.ts
export interface EnumModel {
    id: number;
    value: string;
}

export interface TreatmentEnums {
    treatmentTypes: EnumModel[];
    treatmentStatus: EnumModel[];
}

export interface EmployeeEnums {
  maritalStatus: EnumModel[];
  roles: EnumModel[];
  educationalLevels: EnumModel[];
  genders: EnumModel[];
}